move generated metrics to _metrics_engine and disparities to _disparities
